Remove Selenium and Karma (#13401)
* Remove Selenium from the project [ci skip]

* Remove karma [ci skip]

Note: a huge help was [this](https://github.com/metabase/metabase/pull/9232/commits/8986c457f83dc5d0caacb9d4caf8e030f142707a) unmerged commit by Tom Robinson.

* Remove webchauffeur

* Remove unsupported tests

* Temporarily skip Cypress flake

(blocking the merge of #13401)
