+++ jest.integ.conf.json
@@ -4,17 +4,8 @@
     "\\.(jpg|jpeg|png|gif|eot|otf|webp|svg|ttf|woff|woff2|mp4|webm|wav|mp3|m4a|aac|oga)$": "<rootDir>/frontend/test/__mocks__/fileMock.js",
     "^promise-loader\\?global\\!metabase\\/lib\\/ga-metadata$": "<rootDir>/frontend/src/metabase/lib/ga-metadata.js"
   },
-  "testPathIgnorePatterns": [
-    "<rootDir>/frontend/test/legacy-karma",
-    "<rootDir>/frontend/test/legacy-selenium"
-  ],
-  "testMatch": [
-    "<rootDir>/frontend/test/**/*.integ.spec.js?(x)"
-  ],
-  "modulePaths": [
-    "<rootDir>/frontend/test",
-    "<rootDir>/frontend/src"
-  ],
+  "testMatch": ["<rootDir>/frontend/test/**/*.integ.spec.js?(x)"],
+  "modulePaths": ["<rootDir>/frontend/test", "<rootDir>/frontend/src"],
   "setupFiles": [
     "<rootDir>/frontend/test/jest-setup.js",
     "<rootDir>/frontend/test/enzyme-setup.js",
@@ -28,5 +19,8 @@
   "coverageDirectory": "./",
   "coverageReporters": ["text", "json-summary"],
   "collectCoverageFrom": ["frontend/src/**/*.js", "frontend/src/**/*.jsx"],
-  "coveragePathIgnorePatterns": ["/node_modules/", "/frontend/src/metabase/visualizations/lib/errors.js"]
+  "coveragePathIgnorePatterns": [
+    "/node_modules/",
+    "/frontend/src/metabase/visualizations/lib/errors.js"
+  ]
 }
