+++ package.json
@@ -149,12 +149,6 @@
     "jest-localstorage-mock": "^2.2.0",
     "jscodeshift": "0.5.0",
     "jsonwebtoken": "^7.2.1",
-    "karma": "^1.3.0",
-    "karma-chrome-launcher": "^2.0.0",
-    "karma-jasmine": "^1.1.0",
-    "karma-junit-reporter": "^1.1.0",
-    "karma-nyan-reporter": "^0.2.2",
-    "karma-webpack": "^2.0.5",
     "lint-staged": "^3.3.1",
     "mockdate": "^2.0.2",
     "postcss-cssnext": "^2.4.0",
@@ -167,11 +161,9 @@
     "raf": "^3.4.0",
     "react-test-renderer": "15",
     "sauce-connect-launcher": "^1.1.1",
-    "selenium-webdriver": "^2.53.3",
     "style-loader": "^0.19.0",
     "uglifyjs-webpack-plugin": "^1.0.0",
     "unused-files-webpack-plugin": "^3.0.0",
-    "webchauffeur": "^1.2.0",
     "webpack": "^3.8.1",
     "webpack-dev-server": "^2.9.1",
     "webpack-notifier": "^1.8.0",
@@ -185,7 +177,7 @@
     "lint-prettier": "yarn && prettier -l 'frontend/**/*.{js,jsx,css}' || (echo '\nThese files are not formatted correctly. Did you forget to \"yarn prettier\"?' && false)",
     "lint-flow": "yarn && flow check",
     "lint-docs-links": "yarn && ./bin/verify-doc-links",
-    "test": "yarn test-unit && yarn test-timezones && yarn test-integration && yarn test-cypress && yarn test-e2e && yarn test-karma",
+    "test": "yarn test-unit && yarn test-timezones && yarn test-integration && yarn test-cypress && yarn test-e2e",
     "test-unit": "yarn && jest --maxWorkers=8 --config jest.unit.conf.json",
     "test-unit-watch": "yarn test-unit --watch",
     "test-unit-update-snapshot": "yarn test-unit --updateSnapshot",
@@ -194,8 +186,6 @@
     "test-e2e": "./bin/build-for-test && yarn test-e2e-no-build",
     "test-e2e-watch": "yarn test-e2e --watch",
     "test-e2e-no-build": "yarn && babel-node ./frontend/test/__runner__/run_e2e_tests.js",
-    "test-karma": "yarn && karma start frontend/test/karma.conf.js --single-run",
-    "test-karma-watch": "yarn && karma start frontend/test/karma.conf.js --auto-watch --reporters nyan",
     "test-integrated": "echo 'test-integrated renamed to test-e2e'; exit 1",
     "test-timezones-unit": "yarn && jest --maxWorkers=8 --config jest.tz.unit.conf.json",
     "test-timezones": "yarn && ./frontend/test/__runner__/run_timezone_tests",
